refactor(bedrock): split bedrock plugin implementation by concern
